requirements.txt
aiogram==3.*
python-dotenv>=1.0.0
aiosqlite>=0.19.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
